requirements.txt
annotated-types==0.6.0
anyio==4.3.0
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==21.2.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
bson==0.5.10
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
click==8.0.3
cmdstanpy==1.2.2
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
contourpy==1.2.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.7
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.6.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
exceptiongroup==1.2.1
fastapi==0.110.2
fasteners==0.14.1
filelock==3.13.1
fonttools==4.51.0
future==0.18.2
h11==0.14.0
holidays==0.48
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
importlib_resources==6.4.0
jeepney==0.7.1
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.9.0
minio==7.2.5
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.0
olefile==0.46
packaging==24.0
pandas==2.2.2
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.1.0
prophet==1.1.5
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
pycairo==1.20.1
pycparser==2.22
pycryptodome==3.20.0
pycups==2.0.1
pydantic==2.7.1
pydantic_core==2.18.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==4.7.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.1
stanio==0.5.0
starlette==0.37.2
systemd-python==234
terminator==2.1.1
texttable==1.6.4
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.29.0
virtualenv==20.25.0
wadllib==1.3.6
websocket-client==1.2.3
xdg==5
xkit==0.0.0
zipp==1.0.0
